Voting should be doubles not ints

+++ sn.h
@@ -54,8 +54,8 @@ struct low_res_grid_cell {
 	struct noisy_val n_star, s_star;
 	struct noisy_val n_bar, s_bar;
 	double n, s;
-	int g_star_above, g_star_below;
-	int g_bar_above, g_bar_below;
+	double g_star_above, g_star_below;
+	double g_bar_above, g_bar_below;
 };
 
 void sn_read_from_file(const char *fname, struct sensor_network *sn);
